feat: add authors at footer
