+++ spa/src/components/BackgroundPattern.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="absolute inset-0 opacity-5">
+    <div
+      class="absolute top-0 left-0 w-full h-full bg-gradient-to-br from-blue-500/10 to-purple-500/10"
+    ></div>
+    <div class="absolute top-1/4 right-1/4 w-96 h-96 bg-blue-500/5 rounded-full blur-3xl"></div>
+    <div class="absolute bottom-1/4 left-1/4 w-96 h-96 bg-purple-500/5 rounded-full blur-3xl"></div>
+  </div>
+</template>
